Rollback to last working version using gatekeeper
We (I) failed to get acceptance working using the new
authorization setup (ditching the gatekeeper proxy and using
our authorization middleware).

To ensure acceptance stays available during the holiday period
I'm doing this rollback.
