--- .gatekeeper_env
@@ -1,26 +1,0 @@
-#!/usr/bin/env bash
-
-# Assert this file is SOURCED, not EXECUTED.
-if [ "$0" = "${BASH_SOURCE[0]}" ]; then
-    echo "Error: bash script $0 needs to be sourced, not executed." >&2
-    exit 1
-fi
-
-#export GATEKEEPER_CLIENT_ID=<CLIENT ID>
-#export GATEKEEPER_CLIENT_SECRET=<CLIENT SECRET>
-#export GATEKEEPER_ENCRYPTION_KEY=<ENCRYPTION_KEY>
-
-export GATEKEEPER_LISTEN=0.0.0.0:8000
-export GATEKEEPER_UPSTREAM_URL=http://0.0.0.0:8001
-export GATEKEEPER_DISCOVERY_URL=https://iam.amsterdam.nl/auth/realms/datapunt-acc
-export GATEKEEPER_REDIRECTION_URL=http://blackspots.local:8150
-
-if [ -z "$GATEKEEPER_CLIENT_ID" ]; then
-      echo "\$GATEKEEPER_CLIENT_ID is not set"
-fi
-if [ -z "$GATEKEEPER_CLIENT_SECRET" ]; then
-      echo "\$GATEKEEPER_CLIENT_SECRET is not set"
-fi
-if [ -z "$GATEKEEPER_ENCRYPTION_KEY" ]; then
-      echo "\$GATEKEEPER_ENCRYPTION_KEY is not set"
-fi
