+++ .gitignore
@@ -8,5 +8,4 @@ __pycache__
 *.pytest*
 .pytest_cache
 *.pytest_cache*
-.gatekeeper_env
 .local
