+++ docker-compose.override.yml.sample
@@ -5,7 +5,6 @@ services:
     user: 1000:1000
     ports:
       - 8000:8000
-      - 8001:8001
 
   dev:
     <<: *app
