+++ docker-compose.yml
@@ -18,7 +18,6 @@ services:
       target: app
     ports:
     - 8000
-    - 8001
     volumes:
       - ./src:/src
       - ./tests:/tests
@@ -29,14 +28,6 @@ services:
       DATAPUNT_API_URL: "https://api.data.amsterdam.nl/"
       SECRET_KEY: "dev"
       OBJECTSTORE_PASSWORD: "insecure"
-      # Gatekeeper listens on port 8000 and serves as a proxy for UWSGI
-      PROXY_CLIENT_ID: "${GATEKEEPER_CLIENT_ID}"
-      PROXY_CLIENT_SECRET: "${GATEKEEPER_CLIENT_SECRET}"
-      PROXY_ENCRYPTION_KEY: "${GATEKEEPER_ENCRYPTION_KEY}"
-      PROXY_LISTEN: "${GATEKEEPER_LISTEN}"
-      PROXY_UPSTREAM_URL: "${GATEKEEPER_UPSTREAM_URL}"
-      PROXY_DISCOVERY_URL: "${GATEKEEPER_DISCOVERY_URL}"
-      PROXY_REDIRECTION_URL: "${GATEKEEPER_REDIRECTION_URL}"
       OBJECTSTORE_UPLOAD_CONTAINER_NAME: dev
 
   app:
