+++ requirements.in
@@ -18,3 +18,4 @@ sentry-sdk
 drf_yasg
 django-choices
 datapunt-objectstore
+datapunt-authorization-django
