+++ requirements.txt
@@ -6,9 +6,12 @@
 #
 attrs==20.3.0             # via jsonschema
 certifi==2020.12.5        # via requests, sentry-sdk
+cffi==1.14.4              # via cryptography
 chardet==3.0.4            # via requests
 coreapi==2.3.3            # via django-rest-swagger, drf-yasg, openapi-codec
 coreschema==0.0.4         # via coreapi, drf-yasg
+cryptography==3.3.1       # via jwcrypto
+datapunt-authorization-django==1.1.0  # via -r requirements.in
 datapunt-config-loader==1.1.2  # via datapunt-objectstore
 datapunt-objectstore==2020.9.7  # via -r requirements.in
 debtcollector==2.2.0      # via oslo.config, oslo.utils, python-keystoneclient
@@ -17,7 +20,7 @@ django-choices==1.7.1     # via -r requirements.in
 django-extensions==3.1.0  # via -r requirements.in
 django-filter==2.4.0      # via -r requirements.in
 django-rest-swagger==2.2.0  # via -r requirements.in
-django==2.2.17            # via -r requirements.in, django-choices, django-filter, djangorestframework, drf-yasg
+django==2.2.17            # via -r requirements.in, datapunt-authorization-django, django-choices, django-filter, djangorestframework, drf-yasg
 djangorestframework-csv==2.1.0  # via drf-amsterdam
 djangorestframework-gis==0.16  # via -r requirements.in
 djangorestframework-xml==2.0.0  # via drf-amsterdam
@@ -31,6 +34,7 @@ iso8601==0.1.13           # via keystoneauth1, oslo.utils
 itypes==1.2.0             # via coreapi
 jinja2==2.11.2            # via coreschema
 jsonschema==3.2.0         # via datapunt-config-loader
+jwcrypto==0.8             # via datapunt-authorization-django
 keystoneauth1==4.3.0      # via python-keystoneclient
 markupsafe==1.1.1         # via jinja2
 msgpack==1.0.1            # via oslo.serialization
@@ -45,6 +49,7 @@ oslo.utils==4.7.0         # via oslo.serialization, python-keystoneclient
 packaging==20.8           # via drf-yasg, oslo.utils
 pbr==5.5.1                # via debtcollector, keystoneauth1, os-service-types, oslo.i18n, oslo.serialization, oslo.utils, python-keystoneclient, stevedore
 psycopg2-binary==2.8.6    # via -r requirements.in
+pycparser==2.20           # via cffi
 pyparsing==2.4.7          # via oslo.utils, packaging
 pyrsistent==0.17.3        # via jsonschema
 python-dateutil==2.8.1    # via datapunt-objectstore
@@ -52,13 +57,13 @@ python-keystoneclient==4.2.0  # via -r requirements.in, datapunt-objectstore
 python-swiftclient==3.11.0  # via -r requirements.in, datapunt-objectstore
 pytz==2020.4              # via django, oslo.serialization, oslo.utils
 pyyaml==5.3.1             # via -r requirements.in, datapunt-config-loader, oslo.config
-requests==2.25.0          # via coreapi, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient
+requests==2.25.0          # via coreapi, datapunt-authorization-django, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient
 rfc3986==1.4.0            # via oslo.config
 ruamel.yaml.clib==0.2.2   # via ruamel.yaml
 ruamel.yaml==0.16.12      # via drf-yasg
 sentry-sdk==0.19.5        # via -r requirements.in
 simplejson==3.17.2        # via django-rest-swagger
-six==1.15.0               # via debtcollector, django-choices, djangorestframework-csv, jsonschema, keystoneauth1, oslo.i18n, oslo.utils, python-dateutil, python-keystoneclient, python-swiftclient
+six==1.15.0               # via cryptography, debtcollector, django-choices, djangorestframework-csv, jsonschema, keystoneauth1, oslo.i18n, oslo.utils, python-dateutil, python-keystoneclient, python-swiftclient
 sqlparse==0.4.1           # via django
 stevedore==3.3.0          # via keystoneauth1, oslo.config, python-keystoneclient
 unicodecsv==0.14.1        # via djangorestframework-csv
