+++ requirements_dev.txt
@@ -8,10 +8,13 @@ appdirs==1.4.4            # via virtualenv
 attrs==20.3.0             # via -r ./requirements.txt, jsonschema, pytest
 backcall==0.2.0           # via ipython
 certifi==2020.12.5        # via -r ./requirements.txt, requests, sentry-sdk
+cffi==1.14.4              # via -r ./requirements.txt, cryptography
 chardet==3.0.4            # via -r ./requirements.txt, requests
 coreapi==2.3.3            # via -r ./requirements.txt, django-rest-swagger, drf-yasg, openapi-codec
 coreschema==0.0.4         # via -r ./requirements.txt, coreapi, drf-yasg
 coverage==5.3             # via -r requirements_dev.in, pytest-cov
+cryptography==3.3.1       # via -r ./requirements.txt, jwcrypto
+datapunt-authorization-django==1.1.0  # via -r ./requirements.txt
 datapunt-config-loader==1.1.2  # via -r ./requirements.txt, datapunt-objectstore
 datapunt-objectstore==2020.9.7  # via -r ./requirements.txt
 debtcollector==2.2.0      # via -r ./requirements.txt, oslo.config, oslo.utils, python-keystoneclient
@@ -24,7 +27,7 @@ django-debug-toolbar==3.2  # via -r requirements_dev.in
 django-extensions==3.1.0  # via -r ./requirements.txt
 django-filter==2.4.0      # via -r ./requirements.txt
 django-rest-swagger==2.2.0  # via -r ./requirements.txt
-django==2.2.17            # via -r ./requirements.txt, django-choices, django-cors-headers, django-debug-toolbar, django-filter, djangorestframework, drf-yasg, model-bakery
+django==2.2.17            # via -r ./requirements.txt, datapunt-authorization-django, django-choices, django-cors-headers, django-debug-toolbar, django-filter, djangorestframework, drf-yasg, model-bakery
 djangorestframework-csv==2.1.0  # via -r ./requirements.txt, drf-amsterdam
 djangorestframework-gis==0.16  # via -r ./requirements.txt
 djangorestframework-xml==2.0.0  # via -r ./requirements.txt, drf-amsterdam
@@ -46,6 +49,7 @@ itypes==1.2.0             # via -r ./requirements.txt, coreapi
 jedi==0.17.2              # via ipython
 jinja2==2.11.2            # via -r ./requirements.txt, coreschema
 jsonschema==3.2.0         # via -r ./requirements.txt, datapunt-config-loader
+jwcrypto==0.8             # via -r ./requirements.txt, datapunt-authorization-django
 keystoneauth1==4.3.0      # via -r ./requirements.txt, python-keystoneclient
 markupsafe==1.1.1         # via -r ./requirements.txt, jinja2
 mccabe==0.6.1             # via flake8
@@ -70,6 +74,7 @@ psycopg2-binary==2.8.6    # via -r ./requirements.txt
 ptyprocess==0.6.0         # via pexpect
 py==1.10.0                # via pytest, tox
 pycodestyle==2.6.0        # via flake8
+pycparser==2.20           # via -r ./requirements.txt, cffi
 pyflakes==2.2.0           # via flake8
 pygments==2.7.3           # via ipython
 pyparsing==2.4.7          # via -r ./requirements.txt, oslo.utils, packaging
@@ -82,13 +87,13 @@ python-keystoneclient==4.2.0  # via -r ./requirements.txt, datapunt-objectstore
 python-swiftclient==3.11.0  # via -r ./requirements.txt, datapunt-objectstore
 pytz==2020.4              # via -r ./requirements.txt, django, oslo.serialization, oslo.utils
 pyyaml==5.3.1             # via -r ./requirements.txt, datapunt-config-loader, oslo.config
-requests==2.25.0          # via -r ./requirements.txt, coreapi, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient
+requests==2.25.0          # via -r ./requirements.txt, coreapi, datapunt-authorization-django, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient
 rfc3986==1.4.0            # via -r ./requirements.txt, oslo.config
 ruamel.yaml.clib==0.2.2   # via -r ./requirements.txt, ruamel.yaml
 ruamel.yaml==0.16.12      # via -r ./requirements.txt, drf-yasg
 sentry-sdk==0.19.5        # via -r ./requirements.txt
 simplejson==3.17.2        # via -r ./requirements.txt, django-rest-swagger
-six==1.15.0               # via -r ./requirements.txt, debtcollector, django-choices, djangorestframework-csv, jsonschema, keystoneauth1, oslo.i18n, oslo.utils, python-dateutil, python-keystoneclient, python-swiftclient, tox, virtualenv
+six==1.15.0               # via -r ./requirements.txt, cryptography, debtcollector, django-choices, djangorestframework-csv, jsonschema, keystoneauth1, oslo.i18n, oslo.utils, python-dateutil, python-keystoneclient, python-swiftclient, tox, virtualenv
 sqlparse==0.4.1           # via -r ./requirements.txt, django, django-debug-toolbar
 stevedore==3.3.0          # via -r ./requirements.txt, keystoneauth1, oslo.config, python-keystoneclient
 toml==0.10.2              # via pytest, tox
