--- src/gatekeeper.conf
@@ -1,50 +1,0 @@
-upstream-keepalives: false
-enable-encrypted-token: false
-enable-token-header: false
-enable-authorization-header: false
-enable-authorization-cookies: false
-enable-default-deny: false
-enable-refresh-tokens: true
-enable-logging: true
-preserve-host: true
-secure-cookie: true # set this to false for local dev
-enable-logging: true
-verbose: true
-resources:
-- uri: /*
-  methods:
-  - DELETE
-  - GET
-  - HEAD
-  - PATCH
-  - POST
-  - PUT
-  - TRACE
-  roles:
-  - bs_all
-- uri: /*
-  methods:
-  - GET
-  - HEAD
-  - TRACE
-  roles:
-  - bs_r
-- uri: /status/*
-  white-listed: true
-- uri: /blackspots/redoc/
-  white-listed: true
-- uri: /blackspots/swagger.yaml
-  white-listed: true
-- uri: /favicon.ico
-  white-listed: true
-cors-origins:
-- 'wbakaart.amsterdam.nl'
-- 'acc.wbakaart.amsterdam.nl'
-cors-methods:
-- DELETE
-- GET
-- HEAD
-- PATCH
-- POST
-- PUT
-- TRACE
