+++ src/uwsgi.ini
@@ -7,7 +7,7 @@ vacuum = true
 processes = 4
 threads = 2
 
-http = :8001
+http = :8000
 wsgi-file = main/wsgi.py
 route = /blackspots/static/(.*) static:/src/static/$1
